+++ lib/src/models/code_issue_severity.dart
@@ -1,1 +1,5 @@
-enum CodeIssueSeverity { style }
+enum CodeIssueSeverity {
+  style,
+  warning,
+  error,
+}