+++ lib/src/rules_factory.dart
@@ -1,3 +1,4 @@
+import 'rules/avoid_preserve_whitespace_false.dart';
 import 'rules/base_rule.dart';
 import 'rules/double_literal_format_rule.dart';
 import 'rules/newline_before_return.dart';
@@ -5,6 +6,7 @@ import 'rules/no_boolean_literal_compare_rule.dart';
 import 'rules/no_empty_block.dart';
 
 const _implementedRules = {
+  AvoidPreserveWhitespaceFalseRule(),
   DoubleLiteralFormatRule(),
   NewlineBeforeReturnRule(),
   NoBooleanLiteralCompareRule(),