+++ test/rules/avoid_preserve_whitespace_false_test.dart
@@ -1,0 +1,62 @@
+@TestOn('vm')
+import 'package:analyzer/dart/analysis/features.dart';
+import 'package:analyzer/dart/analysis/utilities.dart';
+import 'package:dart_code_metrics/src/models/code_issue_severity.dart';
+import 'package:dart_code_metrics/src/rules/avoid_preserve_whitespace_false.dart';
+import 'package:test/test.dart';
+
+const _content = '''
+@Component(
+  selector: 'component-selector',
+  templateUrl: 'component.html',
+  styleUrls: ['component.css'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  preserveWhitespace: false,
+  directives: <Object>[
+    coreDirectives,
+  ],
+)
+class Component {}
+
+@Component(
+  selector: 'component2-selector',
+  templateUrl: 'component2.html',
+  styleUrls: ['component2.css'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  directives: <Object>[
+    coreDirectives,
+  ],
+)
+class Component2 {}
+''';
+
+void main() {
+  test('AvoidPreserveWhitespaceFalseRule reports about found issues', () {
+    final sourceUrl = Uri.parse('/example.dart');
+
+    final parseResult = parseString(
+        content: _content,
+        featureSet: FeatureSet.fromEnableFlags([]),
+        throwIfDiagnostics: false);
+
+    final issues = const AvoidPreserveWhitespaceFalseRule()
+        .check(parseResult.unit, sourceUrl, parseResult.content);
+
+    expect(issues.length, equals(1));
+
+    final issue = issues.first;
+    expect(issue.ruleId, equals('avoid-preserve-whitespace-false'));
+    expect(issue.severity, equals(CodeIssueSeverity.warning));
+    expect(issue.sourceSpan.sourceUrl, equals(sourceUrl));
+    expect(issue.sourceSpan.start.offset,  equals(164));
+    expect(issue.sourceSpan.start.line, equals(6));
+    expect(issue.sourceSpan.start.column, equals(3));
+    expect(issue.sourceSpan.end.offset, equals(189));
+    expect(issue.sourceSpan.end.line, equals(6));
+    expect(issue.sourceSpan.end.column, equals(28));
+    expect(issue.sourceSpan.text, equals('preserveWhitespace: false'));
+    expect(issue.message, equals('Avoid use preserveWhitespace: false.'));
+    expect(issue.correction, isNull);
+    expect(issue.correctionComment, isNull);
+  });
+}